\docType{package}
\name{runExamplesCheck}
\alias{runExamplesCheck-package}
\title{Demonstrates the use of macro 'runExamples'}
\description{
  Demonstrates the use of macro runExamples.

  The examples section on this page is generated by putting
  \code{\\runExamples{2+2; pi; cos(pi)}} at the top level of the Rd file
  generating this page.

  R's checking tools (up to version 3.5.1)  give the warning:
\preformatted{  
  prepare_Rd: runExamplesCheck-package.Rd:21:
      Section \Sexpr is unrecognized and will be dropped
}

  Run the checks with to see if a current version of R still gives the warning.
  R-devel (2018-10-02 r75388) does not.
}
\runExamples{2+2; pi; cos(pi)}
